+++ docs/scripts/views/components/alert.vue
@@ -2,21 +2,21 @@
   <ui-page-structure name="alert" demoCount="1" withoutCss>
     <template #hero>
       <div class="hero-demo">
-        <ui-alert :icon="icon" :iconOutlined="iconOutlined">{{
+        <ui-alert :state="state" :stateOutlined="stateOutlined">{{
           message
         }}</ui-alert>
       </div>
       <div class="hero-options">
         <div class="hero-option hero-options">
-          <ui-select :options="IconTypeOptions" v-model="icon"
-            >Icon Type</ui-select
+          <ui-select :options="StateTypeOptions" v-model="state"
+            >State Type</ui-select
           >
           <ui-form-field>
             <ui-checkbox
-              id="icon-outlined"
-              v-model="iconOutlined"
+              id="state-outlined"
+              v-model="stateOutlined"
             ></ui-checkbox>
-            <label for="icon-outlined">Outlined Icon</label>
+            <label for="state-outlined">Outlined State</label>
           </ui-form-field>
         </div>
         <ui-textfield class="hero-option" v-model="message">
@@ -27,17 +27,17 @@
 
     <!-- Content -->
     <section class="example">
-      <ui-alert icon="success">Success Message</ui-alert>
-      <ui-alert icon="info">Info Message</ui-alert>
-      <ui-alert icon="warning" closable>Warning Message</ui-alert>
-      <ui-alert icon="error">Error Message</ui-alert>
+      <ui-alert state="success">Success Message</ui-alert>
+      <ui-alert state="info">Info Message</ui-alert>
+      <ui-alert state="warning" closable>Warning Message</ui-alert>
+      <ui-alert state="error">Error Message</ui-alert>
     </section>
     <ui-snippet :code="$store.demos[1]"></ui-snippet>
   </ui-page-structure>
 </template>
 
 <script>
-const IconTypeOptions = [
+const StateTypeOptions = [
   {
     label: 'Success',
     value: 'success'
@@ -66,9 +66,9 @@ export default {
   },
   data() {
     return {
-      IconTypeOptions,
-      icon: 'success',
-      iconOutlined: false,
+      StateTypeOptions,
+      state: 'success',
+      stateOutlined: false,
       message: 'Message'
     };
   }