+++ docs/scripts/views/plugins/confirm-dialog.vue
@@ -37,7 +37,7 @@ export default {
     showConfrim() {
       this.$confirm({
         message: 'Do you like BalmJS?',
-        icon: 'help',
+        state: 'help',
         acceptText: 'Cool',
         cancelText: 'Good'
       }).then((result) => {